+++ apps/web/src/routes/book-room/-component/split-view/index.tsx
@@ -1,5 +1,5 @@
 export { default as SplitViewGroup } from "./group";
 export { default as SplitViewItem } from "./item";
-export { default as SplitViewHandler } from "./handler";
+export { default as SplitViewSeparator } from "./separator";
 export * from "./context";
 export * from "./hooks/use-split-view";
